# backend/openapi/schema.yaml
openapi: 3.0.3
info:
  title: Chemistry Apps API
  version: 1.0.0
  description: API de plataforma científica modular para ejecutar jobs asíncronos
    por plugins (calculator, random-numbers y molar-fractions), con observabilidad
    de progreso y logs en tiempo real, cache por hash y recuperación activa automática.
  contact:
    name: Chemistry Apps Team
    url: https://github.com/chem-gl/chemistry-apps
  license:
    name: MIT
    url: https://opensource.org/license/mit/
paths:
...
  /api/jobs/:
    get:
      operationId: jobs_list
      description: Devuelve el listado de jobs científicos registrados en todas las
        apps, incluyendo su estado actual. Permite filtros opcionales por plugin y
        estado.
      summary: Listar Jobs Científicos
      parameters:
      - in: query
        name: plugin_name
        schema:
          type: string
        description: Filtra por nombre de plugin/app científica.
      - in: query
        name: status
        schema:
          type: string
        description: 'Filtra por estado del job. Valores válidos: pending, running,
          paused, completed, failed.'
      tags:
      - Jobs
      security:
      - cookieAuth: []
      - basicAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ScientificJob'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                EstadoInválido:
                  value:
                    detail: 'Invalid status filter. Allowed values are: pending, running,
                      paused, completed, failed.'
                  summary: Estado inválido
          description: Filtro de estado inválido.
    post:
      operationId: jobs_create
      description: Evalua reglas de hashing/caching y despacha al worker Celery solo
        cuando el job no tiene cache disponible.
      summary: Despachar Job Científico
      tags:
      - Jobs
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JobCreateRequest'
            examples:
              EjemploDeCreaciónDeJob:
                value:
                  plugin_name: calculator
                  version: '1.0'
                  parameters:
                    op: add
                    a: 5
                    b: 3
                summary: Ejemplo de Creación de Job
                description: Creación de una tarea matemática en background.
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/JobCreateRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/JobCreateRequest'
        required: true
      security:
      - cookieAuth: []
      - basicAuth: []
      - {}
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScientificJob'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                ErrorPluginNoRegistrado:
                  value:
                    detail: Plugin requested is not registered
                  summary: Error plugin no registrado
          description: Error de validacion en parametros.
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      description: Schema consistente para respuestas de error API.
      properties:
        detail:
          type: string
          description: Descripción del error manejado por API.
      required:
      - detail
...
    JobCreateRequest:
      type: object
      description: |-
        Serializer puro para las peticiones de creación de un nuevo
        Job científico, siguiendo fuertemente el schema de OpenAPI.
      properties:
        plugin_name:
          type: string
          minLength: 1
          description: El nombre registrado del plugin o algoritmo a ejecutar.
          maxLength: 100
        version:
          type: string
          minLength: 1
          default: '1.0'
          description: Versión asociada del algoritmo que será parte de la llave de
            caché.
          maxLength: 50
        parameters:
          description: Diccionario dinámico de inputs o parámetros del job.
      required:
      - parameters
      - plugin_name
...
    ScientificJob:
      type: object
      description: |-
        Serializer central que describe toda la entidad del Job.
        Usado como schema de salida.
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        job_hash:
          type: string
          readOnly: true
          description: Hash SHA-256 for caching
        plugin_name:
          type: string
          readOnly: true
        algorithm_version:
          type: string
          readOnly: true
        status:
          allOf:
          - $ref: '#/components/schemas/StatusEnum'
          readOnly: true
        cache_hit:
          type: boolean
          readOnly: true
        cache_miss:
          type: boolean
          readOnly: true
        progress_percentage:
          type: integer
          readOnly: true
          description: Porcentaje de progreso entre 0 y 100.
        progress_stage:
          type: string
          readOnly: true
          description: 'Etapa actual de ejecución: pending/queued/running/paused/caching/completed/failed.'
        progress_message:
          type: string
          readOnly: true
          description: Mensaje corto y legible del estado de progreso.
        progress_event_index:
          type: integer
          readOnly: true
          description: Contador incremental de eventos de progreso emitidos.
        supports_pause_resume:
          type: boolean
          readOnly: true
          description: Indica si el plugin permite pausa cooperativa y reanudación.
        pause_requested:
          type: boolean
          readOnly: true
          description: Marca de control cooperativo para solicitar pausa de ejecución.
        runtime_state:
          readOnly: true
          description: Estado serializable de ejecución para reanudar tareas pausadas.
        paused_at:
          type: string
          format: date-time
          readOnly: true
          nullable: true
          description: Marca temporal del último momento en que el job quedó en pausa.
        resumed_at:
          type: string
          format: date-time
          readOnly: true
          nullable: true
          description: Marca temporal de la última reanudación explícita del job.
        parameters:
          readOnly: true
        results:
          readOnly: true
          nullable: true
        error_trace:
          type: string
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - algorithm_version
      - cache_hit
      - cache_miss
      - created_at
      - error_trace
      - id
      - job_hash
      - parameters
      - pause_requested
      - paused_at
      - plugin_name
      - progress_event_index
      - progress_message
      - progress_percentage
      - progress_stage
      - results
      - resumed_at
      - runtime_state
      - status
      - supports_pause_resume
      - updated_at
...
    StatusEnum:
      enum:
      - pending
      - running
      - paused
      - completed
      - failed
      - cancelled
      type: string
      description: |-
        * `pending` - Pending
        * `running` - Running
        * `paused` - Paused
        * `completed` - Completed
        * `failed` - Failed
        * `cancelled` - Cancelled
...
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid
servers:
- url: http://localhost:8000
  description: Entorno local por hostname
- url: http://127.0.0.1:8000
  description: Entorno local por loopback
...